users must now be logged in to update or delete their own comments

+++ backend/controllers/comments.js
@@ -4,49 +4,54 @@ const User = require('../models/user');
 const Listing = require('../models/listing');
 const jwt = require('jsonwebtoken');
 
-// commentsRouter.get('/', async (req, res) => { // not needed
-//   const comments = await Comment.find({}).populate('user', { username: 1, name: 1 });
-//   res.json(comments);
-// });
-
 const getTokenFrom = req => {
-  const auth = req.get('authorization');
-  if (auth && auth.startsWith('Bearer ')) {
-    return auth.replace('Bearer ', '');
-  }
-  return null;
+	const auth = req.get('authorization');
+	if (auth && auth.startsWith('Bearer ')) {
+		return auth.replace('Bearer ', '');
+	}
+	return null;
 };
 
 commentsRouter.post('/', async (req, res, next) => {
-  const body = req.body;
+	const body = req.body;
 
+	const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET);
+
+	if (!decodedToken.id) {
+		return res.status(401).json({ error: 'token invalid' });
+	}
+
+	const user = await User.findById(decodedToken.id);
+	const listing = await Listing.findById(body.listingId);
+
+	const comment = new Comment({
+		content: body.content,
+		user: user.id,
+		listing: listing.id
+	});
+
+	const savedComment = await comment.save();
+
+	listing.comments = listing.comments.concat(savedComment._id);
+	await listing.save();
+
+	res.json(savedComment);
+});
+
+commentsRouter.put('/:id', async (req, res, next) => {
   const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET);
 
   if (!decodedToken.id) {
     return res.status(401).json({ error: 'token invalid' });
   }
 
+  const comment = await Comment.findById(req.params.id);
   const user = await User.findById(decodedToken.id);
-  const listing = await Listing.findById(body.listingId);
 
-  const comment = new Comment({
-    content: body.content,
-    user: user.id,
-    listing: listing.id
-  });
+  if (comment.user.toString() !== user.id) {
+    return res.status(401).json({ error: 'unauthorized user'});
+  }
 
-  const savedComment = await comment.save();
-
-  // user.comments = user.comments.concat(savedComment._id);
-  // await user.save();
-
-  listing.comments = listing.comments.concat(savedComment._id);
-  await listing.save();
-
-  res.json(savedComment);
-});
-
-commentsRouter.put('/:id', async (req, res, next) => {
   Comment.findByIdAndUpdate(req.params.id, req.body, { new: true })
     .then(updatedComment => {
       res.json(updatedComment);
@@ -54,18 +59,27 @@ commentsRouter.put('/:id', async (req, res, next) => {
     .catch(error => next(error));
 });
 
-commentsRouter.delete('/:id', async (req, res, next) => { // in body: listingId, userId
+commentsRouter.delete('/:id', async (req, res, next) => { // in body: listingId
+  const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET);
+  
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'token invalid' });
+  }
+
   const comment = await Comment.findById(req.params.id);
-  const user = await User.findById(req.body.userId);
+  const user = await User.findById(decodedToken.id);
   const listing = await Listing.findById(req.body.listingId);
 
-  user.comments = user.comments.filter(c => c._id.toString() !== comment._id.toString());
+  if (comment.user.toString() !== user.id) {
+    return res.status(401).json({ error: 'unauthorized user' });
+  }
+
   listing.comments = listing.comments.filter(l => l._id.toString() !== comment._id.toString());
 
-  await user.save();
   await listing.save();
-
   await Comment.findByIdAndRemove(req.params.id);
+
+  res.status(204).end();
 });
 
 module.exports = commentsRouter;
